Added additional unsupported dtypes in cummax

+++ ivy/functional/backends/paddle/statistical.py
@@ -349,7 +349,7 @@ def cumsum(
 
 
 @with_unsupported_device_and_dtypes(
-    {"2.4.2 and below": {"cpu": ("uint16", "bfloat16", "int8",
+    {"2.4.2 and below": {"cpu": ("uint8", "bfloat16", "int8",
                                  "int16", "uint16", "float16")}}, backend_version
 )
 def cummax(
